started moving commands from robot container to dedicated commands. Also pathplanner prelim stuff, + new auto PID consts
